+++ src/imio/helpers/cache.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from datetime import datetime
+from persistent.mapping import PersistentMapping
 from plone import api
 from plone.i18n.normalizer import IIDNormalizer
 from plone.memoize import ram
@@ -9,7 +10,6 @@ from zope.component import getAllUtilitiesRegisteredFor
 from zope.component import getUtility
 from zope.component import queryUtility
 from zope.schema.interfaces import IVocabularyFactory
-from persistent.mapping import PersistentMapping
 
 import logging